+++ cli/src/terminal.rs
@@ -182,7 +182,10 @@ pub fn print_cert_info(
     };
     let cert = &cert.cert;
 
-    let fingerprint = Sha256::digest(cert.to_der().unwrap());
+    let fingerprint = Sha256::digest(
+        cert.to_der()
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?,
+    );
     let slot_id: u8 = slot.into();
     print_cert_attr(stream, "Slot", format!("{:x}", slot_id))?;
     print_cert_attr(